Filling out the React style Buttplug sidebar
This needs to connect to some form of device controller, and isn't hooked up
to Redux yet. Once a controller widget exists, I believe it should use Redux
to comunicate with the sidepanel (which contains the Buttplug Client).

After that, the task will be to relay the device list and control over
the Jitsi message bus between users.